feat: Extract existing columns to reusable constants
